Extract handling clicks to class, use DI
